requirements.txt
flask>=2.0.0
flask-session>=0.4.0
langchain-google-genai>=0.0.2
python-dotenv>=0.19.0